+++ tests/test_component_tags.py
@@ -9,14 +9,14 @@ import pytest
 
 def test_basic_component(request):
     """Tests a simple component with / end"""
-    content = render_component(request, "{% @ main.default.simple_basic_component / %}")
+    content = render_component(request, "{% @ main.default.SimpleBasicComponent / %}")
     assert extract_component(content) == "foo"
 
 
 def test_basic_component_with_end_tag(request):
     """Tests a simple component with  /@ end tag"""
     content = render_component(
-        request, "{% @ main.default.simple_basic_component %}{% /@ %}"
+        request, "{% @ main.default.SimpleBasicComponent %}{% /@ %}"
     )
     assert extract_component(content) == "foo"
 
@@ -25,7 +25,7 @@ def test_basic_component_with_end_tag_and_name(request):
     """Tests a simple component with `/@ <name>` end tag"""
     content = render_component(
         request,
-        "{% @ main.default.simple_basic_component %}{% /@ simple_basic_component%}",
+        "{% @ main.default.SimpleBasicComponent %}{% /@ SimpleBasicComponent %}",
     )
     assert extract_component(content) == "foo"
 
@@ -33,9 +33,9 @@ def test_basic_component_with_end_tag_and_name(request):
 def test_basic_component_with_missing_end_tag(request):
     """Tests a simple component without end tag - must produce TemplateSyntaxError"""
     with pytest.raises(TemplateSyntaxError):
-        content = render_component(
+        render_component(
             request,
-            "{% @ main.default.simple_basic_component %}",
+            "{% @ main.default.SimpleBasicComponent %}",
         )